# Prevent unnecessary memory allocations in WriteIPFIXMsgToBuffer