@antoninbas 2025-01-09
Before change:
```
C02F16ASMD6R:go-ipfix abas$ go test -run=xxx -bench=BenchmarkWriteIPFIXMsgToBuffer -benchmem ./pkg/exporter/
goos: darwin
goarch: amd64
pkg: github.com/vmware/go-ipfix/pkg/exporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkWriteIPFIXMsgToBuffer-16    	  291412	      3545 ns/op	      80 B/op	      10 allocs/op
PASS
ok  	github.com/vmware/go-ipfix/pkg/exporter	51.294s
```

After change:
```
C02F16ASMD6R:go-ipfix abas$ go test -run=xxx -bench=BenchmarkWriteIPFIXMsgToBuffer -benchmem ./pkg/exporter/
goos: darwin
goarch: amd64
pkg: github.com/vmware/go-ipfix/pkg/exporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkWriteIPFIXMsgToBuffer-16    	  378382	      3050 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/vmware/go-ipfix/pkg/exporter	37.357s
```